test(runtime): isolate bare throw default regression
